Relocate deferred prompt to prompt utils
