+++ types/src/prompts/ddl/ui.d.ts
@@ -1,17 +1,18 @@
 import { Component } from 'preact';
-import { DdlPromptConfig } from '../../core';
+import { DDLPromptConfig } from '../../core';
 interface UiProps {
-    config: DdlPromptConfig;
-    onBannerConfirm: (config: DdlPromptConfig) => void;
-    onBannerCancelled: (config: DdlPromptConfig) => void;
+    config?: DDLPromptConfig;
+    onBannerConfirm: (config: DDLPromptConfig) => void;
+    onBannerCancelled: (config: DDLPromptConfig) => void;
 }
 export default class Ui extends Component<UiProps, never> {
     private siteMargin?;
     private siteTransition?;
     onDimensions: (_bannerWidth: number, bannerHeight: number) => void;
-    onBannerConfirm: (config: DdlPromptConfig) => void;
-    onBannerCancelled: (config: DdlPromptConfig) => void;
+    onBannerConfirm: (config: any) => void;
+    onBannerCancelled: (config: any) => void;
     resetBodyStyles(): void;
+    renderBanner(): void;
     render(): import("preact").VNode<any> | null;
 }
 export {};
