+++ types/src/prompts/utils.d.ts
@@ -1,0 +1,2 @@
+import { PromptConfig } from '../core';
+export declare function deferPromptActivation(prompt: PromptConfig, activateFn: (prompt: PromptConfig) => void): boolean;
